requirements.txt
Django==3.1.7
configparser==5.0.2
PyMySQL==1.0.2
django-picklefield==3.0.1
